third_party/WebKit/LayoutTests/fast/events/touch/gesture/pad-gesture-cancel.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../../resources/js-test.js"></script>
<script>
    description("Tests basic use of GestureFlingCancel");

    var expectedWheelEventsOccurred = "0";
    var actualWheelEventsOccurred = 0;

    function notifyWhenFlingIsOver() {
        if (!window.testRunner)
            return;
        if (eventSender.isFlinging()) {
            window.setTimeout(notifyWhenFlingIsOver, 0);
            return;
        }
        shouldBe('actualWheelEventsOccurred', expectedWheelEventsOccurred);
        testRunner.notifyDone();
    }

    function recordWheelEvent(event)
    {
        shouldBe('event.clientX', "10");
        shouldBe('event.clientY', "11");

        // Test deliberately does not equality check wheelDeltas to not be fragile in the face of curve adjustment.
        shouldBeTrue("event.wheelDeltaX > 5");
        shouldBeTrue("event.wheelDeltaY > 5");
        actualWheelEventsOccurred++;
        shouldBeTrue("actualWheelEventsOccurred == 0");
        if (window.testRunner)
            testRunner.notifyDone();
    }

    if (window.testRunner && window.eventSender && eventSender.gestureFlingStart) {
        eventSender.gestureFlingStart(10, 11, 1000, 1000, "touchpad");
        eventSender.gestureFlingCancel();
        document.addEventListener("mousewheel", recordWheelEvent);
        testRunner.waitUntilDone();
        notifyWhenFlingIsOver();
    }
</script>
</head>
<body>
<span id="e"></span>
<span id="f"></span>
</body>
</html>
